feat: Add Partner Products

+++ database/migrations/2025_09_26_090000_create_shifts_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->text('notes')->nullable();
 
             // Audit trail fields
-            $table->unsignedBigInteger('created_by')->nullable();
-            $table->unsignedBigInteger('updated_by')->nullable();
+            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedInteger('updated_by')->nullable();
 
             $table->timestamps();
